Change access control to occupations

+++ src/main/java/org/fenixedu/spaces/ui/services/OccupationService.java
@@ -22,6 +22,7 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -31,6 +32,7 @@ import org.fenixedu.bennu.core.security.Authenticate;
 import org.fenixedu.commons.i18n.I18N;
 import org.fenixedu.spaces.core.service.NotificationService;
 import org.fenixedu.spaces.domain.Space;
+import org.fenixedu.spaces.domain.SpaceDomainException;
 import org.fenixedu.spaces.domain.occupation.Occupation;
 import org.fenixedu.spaces.domain.occupation.config.ExplicitConfigWithSettings;
 import org.fenixedu.spaces.domain.occupation.config.ExplicitConfigWithSettings.Frequency;
@@ -166,8 +168,8 @@ public class OccupationService {
 
     @Atomic
     public void createOccupation(String emails, String subject, String description, String selectedSpaces, String config,
-            String events, OccupationRequest request) throws Exception {
-        final Set<Space> selectedSpaceSet = selectSpaces(selectedSpaces);
+            String events, OccupationRequest request, User user) throws Exception {
+        final Set<Space> selectedSpaceSet = selectSpaces(selectedSpaces, user);
         final List<Interval> intervals = selectEvents(events);
         final Occupation occupation = new Occupation(emails, subject, description, parseConfig(config, intervals));
         for (Space space : selectedSpaceSet) {
@@ -244,25 +246,30 @@ public class OccupationService {
         return intervals;
     }
 
-    private Set<Space> selectSpaces(String selectedSpaces) {
+    private Set<Space> selectSpaces(String selectedSpaces, User user) {
         final Set<Space> selectedSpaceSet = new HashSet<>();
         final JsonArray spacesJson = jsonParser.parse(selectedSpaces).getAsJsonArray();
         for (JsonElement spaceJson : spacesJson) {
             String spaceId = spaceJson.getAsString();
             final Space space = FenixFramework.getDomainObject(spaceId);
             if (FenixFramework.isDomainObjectValid(space)) {
+                if (!space.isOccupationMember(user)) {
+                    throw new SpaceDomainException("unauthorized.selected.space", space.getName());
+                }
                 selectedSpaceSet.add(space);
             }
         }
         return selectedSpaceSet;
     }
 
-    public List<Occupation> getOccupations(Integer month, Integer year) {
+    public List<Occupation> getOccupations(Integer month, Integer year, User user) {
         DateTime start = new DateTime(year, month, 1, 0, 0);
         Interval interval = new Interval(start, start.plusMonths(1));
+        Predicate<Occupation> userPredicate =
+                o -> o.getSpaces().stream().filter(s -> s.isOccupationMember(user)).findAny().isPresent();
         return Bennu.getInstance().getOccupationSet().stream().filter(o -> o.getClass().equals(Occupation.class))
-                .filter(o -> o.overlaps(interval)).sorted((o1, o2) -> o2.getStart().compareTo(o1.getStart()))
-                .collect(Collectors.toList());
+                .filter(userPredicate).filter(o -> o.overlaps(interval))
+                .sorted((o1, o2) -> o2.getStart().compareTo(o1.getStart())).collect(Collectors.toList());
     }
 
     public String exportConfig(Occupation occupation) {
@@ -313,12 +320,15 @@ public class OccupationService {
     }
 
     @Atomic
-    public void editOccupation(Occupation occupation, String emails, String subject, String description, String selectedSpaces)
-            throws Exception {
+    public void editOccupation(Occupation occupation, String emails, String subject, String description, String selectedSpaces,
+            User user) throws Exception {
+        if (!canManageOccupation(occupation, user)) {
+            throw new SpaceDomainException("unauthorized.edit.occupation");
+        }
         occupation.setEmails(emails);
         occupation.setSubject(subject);
         occupation.setDescription(description);
-        final Set<Space> selectedSpaceSet = selectSpaces(selectedSpaces);
+        final Set<Space> selectedSpaceSet = selectSpaces(selectedSpaces, user);
         occupation.getSpaces().stream().forEach(s -> occupation.removeSpace(s));
         for (Space space : selectedSpaceSet) {
             if (!space.isFree(occupation.getIntervals())) {
@@ -329,8 +339,20 @@ public class OccupationService {
         }
     }
 
+    public boolean canManageOccupation(Occupation occupation, User user) {
+        for (Space space : occupation.getSpaces()) {
+            if (!space.isOccupationMember(user)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
     @Atomic
-    public void delete(Occupation occupation) {
+    public void delete(Occupation occupation, User user) {
+        if (!canManageOccupation(occupation, user)) {
+            throw new SpaceDomainException("unauthorized.delete.occupation");
+        }
         occupation.delete();
     }
